+++ hdbconnect_impl/src/a_sync/prepared_statement.rs
@@ -2,6 +2,7 @@ use crate::{
     a_sync::HdbResponse,
     base::{new_am_async, InternalReturnValue, PreparedStatementCore, AM},
     conn::{AmConnCore, CursorHoldability},
+    impl_err,
     protocol::{
         parts::{
             HdbValue, LobFlags, ParameterDescriptors, ParameterRows, ResultSetMetadata, TypeId,
@@ -9,7 +10,7 @@ use crate::{
         MessageType, Part, PartKind, Request, ServerUsage,
     },
     types_impl::lob::async_lob_writer,
-    ConnectionConfiguration, HdbError, HdbResult,
+    usage_err, ConnectionConfiguration, HdbResult,
 };
 use std::sync::Arc;
 
@@ -253,11 +254,11 @@ impl<'a> PreparedStatement {
             if let Some(Part::WriteLobReply(wlr)) = write_lob_reply {
                 let locator_ids = wlr.into_locator_ids();
                 if locator_ids.len() != readers.len() {
-                    return Err(HdbError::UsageDetailed(format!(
+                    return Err(usage_err!(
                         "{} readers are provided, but {} are required",
                         readers.len(),
                         locator_ids.len()
-                    )));
+                    ));
                 }
 
                 for (locator_id, (reader, type_id)) in locator_ids.into_iter().zip(readers) {
@@ -305,7 +306,7 @@ impl<'a> PreparedStatement {
             self.batch.push(input, &self.a_descriptors)?;
             return Ok(());
         }
-        Err(HdbError::Usage(
+        Err(usage_err!(
             "Batch not usable for PreparedStatements without input parameter",
         ))
     }
@@ -327,7 +328,7 @@ impl<'a> PreparedStatement {
                 .push_hdb_values(hdb_values, &self.a_descriptors)?;
             return Ok(());
         }
-        Err(HdbError::Usage(
+        Err(usage_err!(
             "Batch not possible, PreparedStatement has no input parameter",
         ))
     }
@@ -350,7 +351,7 @@ impl<'a> PreparedStatement {
     /// Several variants of `HdbError` can occur.
     pub async fn execute_batch(&mut self) -> HdbResult<HdbResponse> {
         if self.batch.is_empty() && self.a_descriptors.has_in() {
-            return Err(HdbError::Usage("Empty batch cannot be executed"));
+            return Err(usage_err!("Empty batch cannot be executed"));
         }
         let mut batch2 = ParameterRows::new();
         trace!(
@@ -476,7 +477,7 @@ impl<'a> PreparedStatement {
             }
         }
 
-        let statement_id = o_stmt_id.ok_or_else(|| HdbError::Impl("No StatementId received"))?;
+        let statement_id = o_stmt_id.ok_or_else(|| impl_err!("No StatementId received"))?;
         let am_ps_core = new_am_async(PreparedStatementCore {
             am_conn_core,
             statement_id,